+++ release/directives/visibility.directive.d.ts
@@ -1,4 +1,4 @@
-import { EventEmitter, ElementRef, NgZone } from '@angular/core';
+import { EventEmitter, ElementRef, NgZone, OnInit, OnDestroy } from '@angular/core';
 /**
  * Visibility Observer Directive
  *
@@ -10,9 +10,15 @@ import { EventEmitter, ElementRef, NgZone } from '@angular/core';
  * 		</div>
  *
  */
-export declare class VisibilityDirective {
+export declare class VisibilityDirective implements OnInit, OnDestroy {
+    private element;
+    private zone;
     isVisible: boolean;
     visible: EventEmitter<any>;
+    private timeout;
     constructor(element: ElementRef, zone: NgZone);
-    visbilityChange(): void;
+    ngOnInit(): void;
+    ngOnDestroy(): void;
+    onVisibilityChange(): void;
+    runCheck(): void;
 }