+++ release/utils/index.js
@@ -12,7 +12,6 @@ __export(require('./prefixes'));
 __export(require('./scrollbar-width'));
 __export(require('./selection'));
 __export(require('./translate'));
-__export(require('./visibility-observer'));
 __export(require('./throttle'));
 __export(require('./sort'));
 __export(require('./row-height-cache'));