+++ release/utils/visibility-observer.d.ts
@@ -1,2 +1,0 @@
-import { NgZone } from '@angular/core';
-export declare function checkVisibility(element: any, callback: any, zone: NgZone): void;